Transferred some functionality from AVLTree to AVLNode.
Previously, methods search(), insert(), and delete() on each node in the tree would first traverse up the tree to the root, and then start operating. Aside from being unintuitive, this also caused the user of the tree to constantly check whether his reference to the tree is still valid, and dealing with empty trees was flawed. Now, AVLTree is the principal user interface for the structure, while the majority of functionality was transferred to AVLNode. Nodes can still be invalidated, but the tree is always there and at the worst it is just empty.